requirements.txt
streamlit
pandas
#opencv-python-headless
numpy
Pillow
#face-recognition
requests
opencv-contrib-python